+++ src/components/ids-dropdown/ids-dropdown.ts
@@ -1,6 +1,6 @@
 import { customElement, scss } from '../../core/ids-decorators';
 import { attributes } from '../../core/ids-attributes';
-import { stringToBool } from '../../utils/ids-string-utils/ids-string-utils';
+import { stringToBool, escapeRegExp } from '../../utils/ids-string-utils/ids-string-utils';
 import IdsDropdownAttributeMixin from './ids-dropdown-attributes-mixin';
 import IdsEventsMixin from '../../mixins/ids-events-mixin/ids-events-mixin';
 import IdsKeyboardMixin from '../../mixins/ids-keyboard-mixin/ids-keyboard-mixin';
@@ -31,6 +31,7 @@ import type IdsListBox from '../ids-list-box/ids-list-box';
 import type IdsListBoxOption from '../ids-list-box/ids-list-box-option';
 import type IdsTriggerField from '../ids-trigger-field/ids-trigger-field';
 import type IdsIcon from '../ids-icon/ids-icon';
+import type IdsCheckbox from '../ids-checkbox/ids-checkbox';
 import { IdsPopupElementRef } from '../ids-popup/ids-popup-attributes';
 import { getClosestContainerNode } from '../../utils/ids-dom-utils/ids-dom-utils';
 import { IdsDropdownOption, IdsDropdownOptions } from './ids-dropdown-common';
@@ -527,7 +528,6 @@ export default class IdsDropdown extends Base {
       this.dropdownList.setAttribute(attributes.TRIGGER_ELEM, `#${triggerElemId}`);
       this.dropdownList.popup.alignTarget = this.input?.fieldContainer || null;
 
-      this.dropdownList.setAttribute(attributes.TRIGGER_TYPE, 'click');
       this.dropdownList.popupOpenEventsTarget = document.body;
 
       // Configure inner IdsPopup
@@ -779,7 +779,7 @@ export default class IdsDropdown extends Base {
 
     // Select on Tab
     this.listen(['Tab'], this, (e: KeyboardEvent) => {
-      if (!this.dropdownList?.popup?.visible) {
+      if (!this.dropdownList?.popup?.visible || this.#isMultiSelect) {
         return;
       }
 
@@ -824,24 +824,21 @@ export default class IdsDropdown extends Base {
     this.attachKeyboardOpenEvent();
   }
 
-  #attachTypeaheadEvents() {
-    // Handle Key Typeahead
-    this.offEvent('keydownend.dropdown-typeahead');
-    this.onEvent('keydownend.dropdown-typeahead', this, (e: CustomEvent) => {
-      this.#typeAhead(e.detail.keys);
-    });
-  }
-
-  #removeTypeaheadEvents() {
-    this.offEvent('keydownend.dropdown-typeahead');
-  }
-
   /**
    * Establish Internal Keyboard shortcuts
    * @private
    * @returns {object} This API object for chaining
    */
   #attachKeyboardListeners() {
+    this.offEvent('keydownend.dropdown-typeahead');
+    this.onEvent('keydownend.dropdown-typeahead', this, (e: CustomEvent) => {
+      if (this.typeahead) {
+        this.#typeAhead(e.detail.keys);
+      } else {
+        this.#selectMatch(e.detail.keys);
+      }
+    });
+
     // Handle up and down arrow
     this.listen(['ArrowDown', 'ArrowUp'], this, (e: KeyboardEvent) => {
       e.stopPropagation();
@@ -962,6 +959,44 @@ export default class IdsDropdown extends Base {
   }
 
   /**
+   * Find and select the only option by input text provided
+   * @param {string} input the text to find by
+   */
+  #selectMatch(input: string) {
+    const term: string = escapeRegExp(input)?.trim();
+
+    if (!term) return;
+
+    const option = this.options
+      .filter((item: IdsListBoxOption) => !item.hasAttribute(attributes.GROUP_LABEL))
+      .find((item: IdsListBoxOption) => {
+        const regex = new RegExp(`^(${term})`, 'i');
+        const label = this.#isMultiSelect ? item.querySelector<IdsCheckbox>('ids-checkbox')?.label : item.textContent?.trim();
+
+        return label?.match(regex);
+      });
+
+    if (this.dropdownList?.popup?.visible && option) {
+      this.deselectOption(this.selected);
+      this.selectOption(option as IdsListBoxOption);
+
+      return;
+    }
+
+    const value = option?.getAttribute(attributes.VALUE);
+
+    if (value) {
+      if (this.#isMultiSelect) {
+        const filtered: string[] = (this.value as any).filter((item: string) => item !== value);
+
+        (this.value as any) = [...filtered, value];
+      } else {
+        this.value = value;
+      }
+    }
+  }
+
+  /**
    * Helper to replace trigger button icon
    * @param {string} icon ids-icon icon value
    */
@@ -1043,12 +1078,12 @@ export default class IdsDropdown extends Base {
 
   /**
    * Find matches between the input value and the dataset
-   * @param {string | RegExp} inputValue value of the input field
+   * @param {string} inputValue value of the input field
    * @returns {IdsDropdownOptions} containing matched values
    */
-  #findMatches(inputValue: string | RegExp): IdsDropdownOptions {
+  #findMatches(inputValue: string): IdsDropdownOptions {
     return this.#optionsData.reduce((options: Array<IdsDropdownOption>, option: IdsDropdownOption, index: number) => {
-      const regex = new RegExp(inputValue, 'gi');
+      const regex = new RegExp(`(${escapeRegExp(inputValue)})`, 'gi');
 
       if (option.label?.match(regex) && !option.groupLabel) {
         const groupLabelOption = this.#getGroupLabelOption(index);
@@ -1199,11 +1234,9 @@ export default class IdsDropdown extends Base {
 
     if (val) {
       this.setAttribute(attributes.TYPEAHEAD, String(val));
-      this.#attachTypeaheadEvents();
       this.#setOptionsData();
     } else {
       this.removeAttribute(attributes.TYPEAHEAD);
-      this.#removeTypeaheadEvents();
     }
 
     this.container?.classList.toggle('typeahead', val);